+++ lib/src/assistants.dart
@@ -1,0 +1,6 @@
+import 'client/client.dart';
+
+class Assistants {
+  final OpenAIClient _client;
+  Assistants(this._client);
+}