+++ test/model/fine_tune/request/create_fine_tune_test.dart
@@ -6,8 +6,9 @@ void main() {
   group('openai create fine tune test', () {
     test('openai create fine tune test to json', () {
       final fineTune = CreateFineTune(
-              trainingFile: 'trainingFile', model: Babbage002FineModel())
-          .toJson();
+        trainingFile: 'trainingFile',
+        model: Babbage002FineModel(),
+      ).toJson();
 
       expect(fineTune['training_file'], 'trainingFile');
       expect(fineTune['n_epochs'], 4);